Add -z parameter to rpm/deb build

+++ rpms/generate_rpm_package.sh
@@ -18,6 +18,7 @@ REVISION="1"
 TARGET=""
 JOBS="2"
 DEBUG="no"
+BUILD_DOCKER="yes"
 USER_PATH="no"
 SRC="no"
 RPM_AARCH64_BUILDER="rpm_builder_aarch64"
@@ -80,7 +81,9 @@ build_rpm() {
     fi
 
     # Build the Docker image
-    docker build -t ${CONTAINER_NAME} ${DOCKERFILE_PATH} || return 1
+    if [[ ${BUILD_DOCKER} == "yes" ]]; then
+      docker build -t ${CONTAINER_NAME} ${DOCKERFILE_PATH} || return 1
+    fi
 
     # Build the RPM package with a Docker container
     docker run -t --rm -v ${OUTDIR}:/var/local/wazuh:Z \
@@ -173,6 +176,7 @@ help() {
     echo "    -p, --path <path>            [Optional] Installation path for the package. By default: /var."
     echo "    -d, --debug                  [Optional] Build the binaries with debug symbols and create debuginfo packages. By default: no."
     echo "    -c, --checksum <path>        [Optional] Generate checksum on the desired path (by default, if no path is specified it will be generated on the same directory than the package)."
+    echo "    --dont-build-docker      [Optional] Locally built docker image will be used instead of generating a new one."
     echo "    --sources <path>             [Optional] Absolute path containing wazuh source code. This option will use local source code instead of downloading it from GitHub."
     echo "    --packages-branch <branch>   [Optional] Select Git branch or tag from wazuh-packages repository. e.g ${PACKAGES_BRANCH}"
     echo "    --dev                        [Optional] Use the SPECS files stored in the host instead of downloading them from GitHub."
@@ -246,6 +250,10 @@ main() {
             ;;
         "-d"|"--debug")
             DEBUG="yes"
+            shift 1
+            ;;
+        "--dont-build-docker")
+            BUILD_DOCKER="no"
             shift 1
             ;;
         "-c"|"--checksum")
